mrc-hybrid-arch(extract-hybrid-arch-props-move,mechanical_provable): Extract hybrid-arch helpers to configs.hybrid_arch (cut+paste)
